requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2
python-dotenv
dj-database-url
gunicorn
django-filter